ransom-module/backend/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
python-dotenv
openai
